commit 0.52 Updated Teacher functionality

+++ src/main/java/org/example/kihelp_back/task/dto/TaskResponse.java
@@ -18,6 +18,6 @@ public record TaskResponse(
         boolean autoGenerate,
         Instant createdTimeStamp,
         List<Argument> arguments,
-        Integer teacherId
+        Long teacherId
 ) {
 }
