+++ src/main/java/org/example/kihelp_back/teacher/usecase/TeacherDeleteUseCase.java
@@ -1,5 +1,5 @@
 package org.example.kihelp_back.teacher.usecase;
 
 public interface TeacherDeleteUseCase {
-    void deleteTeacher(Integer teacherId);
+    void deleteTeacher(Long teacherId);
 }
